network/requirements.txt
absl-py==1.4.0
airsim==1.8.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
gdown==3.15.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
grpcio==1.54.0
h5py==3.8.0
hydra-core==0.11.3
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
llvmlite==0.33.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.5
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nose==1.3.7
notebook==6.5.4
notebook-shim==0.2.3
numba==0.50.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
omegaconf==1.4.1
open3d==0.10.0.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
plyfile==0.7.4
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-louvain==0.16
pytorch-metric-learning==0.9.99
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
rdflib==6.3.2
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.16.2
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==1.20.0
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
subprocess32==3.5.4
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.1
torch-cluster==1.6.1
torch-geometric==1.7.2
torch-points-kernels==0.6.10
torch-points3d==1.3.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torchaudio==0.13.1
torchnet==0.0.4
torchvision==0.14.1
tornado==4.5.3
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.5.0
urllib3==2.0.2
visdom==0.2.4
wandb==0.8.36
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
zipp==3.15.0
